Silence more type-related warnings, mostly in the package codes. Fix an integer overflow bug in box breaking code.

+++ source/SAMRAI/algs/OuternodeSumTransaction.h
@@ -174,8 +174,8 @@ private:
    hier::Box d_src_node;
    const xfer::RefineClasses::Data** d_refine_data;
    int d_item_id;
-   int d_incoming_bytes;
-   int d_outgoing_bytes;
+   size_t d_incoming_bytes;
+   size_t d_outgoing_bytes;
 
 };
