+++ source/SAMRAI/mesh/VoucherTransitLoad.C
@@ -852,8 +852,8 @@ VoucherTransitLoad::adjustLoad(
          main_bin.getSumLoad() < low_load ? low_load - main_bin.getSumLoad() : 0;
       tbox::plog << "  adjustLoad point_miss=" << point_miss
                  << "  range_miss="
-                 << (range_miss > 0 ? " ":"") // Add space if missed range
-                 << (range_miss > 0.5*d_pparams->getMinBoxSize().getProduct() ? " ":"") // Add space if missed range by a lot
+                 << (range_miss > 0 ? " " : "") // Add space if missed range
+                 << (range_miss > 0.5 * static_cast<double>(d_pparams->getMinBoxSize().getProduct()) ? " " : "") // Add space if missed range by a lot
                  << range_miss
                  << "  " << main_bin.getSumLoad() << '/'
                  << ideal_load << " [" << low_load << ',' << high_load << ']'
