+++ source/SAMRAI/pdat/CellData.h
@@ -291,7 +291,7 @@ public:
     *
     * @pre dynamic_cast<const CellOverlap *>(&overlap) != 0
     */
-   virtual int
+   virtual size_t
    getDataStreamSize(
       const hier::BoxOverlap& overlap) const;
